"Stop mistaken texts" - Patchset #1
This feature adds an optional delay when sending messages in the application. The delay can be changed.
When the countdown is running, it is possible to stop the message before dispatch by clicking on the item in the message list.

I tried to implement this feature with minimal changes to the structure of the application, to minimize potential impacts on the stability of message
sending.

Note: It's been only tested with SMS for now (and I think it only works with SMSes).

Only issue I can think of right now:
It is not possible to delay the first message of a conversation, because the conversation is created when the first message is sent to the framework
service. This is bypassed using a *secret* technique. However, the message displayed on screen is incorrect (it says it's delaying the text, whereas
it's not). The only solution I could come with, which sucks a lot, is to query the conversation from the SMS Content Resolver, and then check if the
first message has the same timestamp, but even this one is hard to implement because the MessageListItem class is not aware of the conversation it's
representing.

@nebkat: Hope the spacing is A-OK, I DID pay attention this time (it hurts my eyes seeing all the whitespaces now that Eclipse is showing them xD)

Change-Id: Ib1663ac3ccc80bffe18c3e06de0d2c9be543f9a4