"Stop mistaken texts"
This feature adds an optional delay when sending messages in the application. The delay can be changed.
When the countdown is running, it is possible to stop the message before dispatch by clicking on the item in the message list.

Note: It's been only tested with SMS messages for now. But using this for MMS messages should be a piece of cake (it's not included because I don't want to commit something without testing it before-hand).

Only issue:
It is not possible to delay the first message of a conversation, because the conversation is created when the first message is sent to the framework service. As a consequence, the countdown will not be used for the first message of a conversation.

Change-Id: I13b65408f7a679651991e4e1e70b963779940b7a